feat:优化权限管理

+++ backend/app/session/models.py
@@ -12,7 +12,7 @@ class ChatSession(Base):
 
     id = Column(String(64), primary_key=True, index=True)
     title = Column(String(255), nullable=False)
-    member_id = Column(Integer, nullable=False, index=True)
+    member_id = Column(BigInteger, nullable=False, index=True)
     create_at = Column(DateTime, server_default=func.now(), nullable=False)
     token_use = Column(Integer, nullable=True)
     session_type = Column(String(64), nullable=False, default=SessionType.CHAT.value)
